Offer the possibility to close the ressource used by the OsmParser.
I added a close() method to the OsmParser Trait and implemented it in the OsmXmlParser and the OsmPbfParser.

The OsmPbfParser will forward this call to it's DataInputStream.

And the OsmXmlParser will forward it to its Source that will forward it to its FileInputStream as can be seen in the source code: https://github.com/scala/scala/blob/v2.11.7/src/library/scala/io/Source.scala#L97